requirements.txt
azure-functions
PyJWT
cryptography

# Development dependencies
pytest
pytest-cov
pytest-mock